+++ CHANGELOG.md
@@ -1,7 +1,21 @@
-# Changelog
+## WIP
 
-### WIP
-- 
+# Changelog
+4/4
+- added ability to couple/decouple selections
+    - checkbox in ui sets a global var IS_COUPLED
+    - can be changed during file analysis without needing to close and resubmit
+    - adjusted onselect in int plot to make onselect functions for widgets which allows the script to read which box was just selected
+    - when coupled data analysis and plot updating is business as usual
+    - when decoupled only one span selector changes, all others are left unchanged
+        - the data from the changed selector is what is analyzed
+    - onselect passes a dict of every columns xmin and xmax from the span selectors
+        - this allows the stats from other columns to be kept while analyzing the selected columns
+    - when uncoupled, each columns xmin/xmax are shown
+    - when coupled one xmin/xmax is shown that applies to all columns
+- added a clear output file button for use when changing data files
+    - this is due to the nature of the coupling feature reading the previous output data to update only one
+- added an exit button
 
 2/6
 - updated readme to reflect solution to error with spyder